refactor(orm): structure the orm pkg
